requirements.txt
config
coverage
jsonpickle
kafka-python==2.0.1